Streamline print, plotting logic for multiple AM1, AM2, AM3.
Gracefully catch new ValidateErrorNoStars for AM1, AM2, AM3.

The astrometric repeatability looks a different distances:
5, 20, and 200 arcminutes.  It quite possible that a small field,
or single chip will not have the larger of these distances.
ValidateErrorNoStars is raised for these cases and caught by the
main validate.py, which then skips further printing, plotting, or saving
of the AMx that was not calculable.

Add `util.calcOrNone` function that catches a specified error
and returns None if that error is raised.  Otherwise returns the
result of the function.  Passes through any other errors.