Implement serialization for PA1, PF1 and PA2
In calcSrd there are serializers that define the metrics as well as
serializers that define parameter sets.

In validate the measurements are captured with a MeasurementSerializer.

This setup can certainly be refactored, streamlined, and made more
robust by turning the measurement functions into classes that implement
an API that consistently provides the measurement itself in addition to
metadata.