Create matchreduce module for data reduction blobs
matchreduce consolidates the functions for matching stars and doing
preliminary analysis against them. matchreduce uses BlobBase so that the
results are serializable.

Also moves all related functions from the check module to this
matchreduce module.

matchreduce contains classes

- MatchedMultiVisitDataset
- AnalyticAstrometryModel
- AnalyticPhotmetryModel

MatchedMultiVisitDataset contains the former loadAndMatchData and
analyzeData functions from validate into a new class that acts both as a
data container for the results of the catalog reduction, and a
persistence layer to JSON.