Making a blob a property links it to measurement
Previously one had to explicitly link a blob to a measurement with the
linkBlob method in measurement classes. This commit uses knowledge that
a blob is a subclass of blob base to automatically add it to the
Measurement class's blob inventory when the blob is set as a instance
attribute.