Change to astropy Quantity-oriented API
Previously validate_base APIs treated astropy.units.Quantity as an
add-on. Values were stored in their original form (as floats, ints or
arrays) with unit information stored separately in a string. While this
meant that we had minimal dependently on astropy itself, it also meant
that we weren't taking advantage of the safety of quantities (for
example, in comparisons. Updating a value wouldn't ensure that the units
were appropriate.

This mass update makes astropy Quantities the core storage for all
numeric values. The main change is that now the API requires that the
user provides quantities. This is done to eliminate confusion. The user
puts quantities into the API. validate_base manages those quantities,
and validate_base provides those quantities when accessed by the user.
This is better than sometimes using raw values and sometimes using
quantities.

In additional to quantities, Datum types also accept strings and
booleans. Unitless values are already supported by astropy Quantity.

Attribute names
---------------

For consistency, some of the Datum and related attributes are renamed:

- `value` is not `quantity`, since `value` is not itself an *attribute*
  of `quantity` representing the unit-less number
- `units` is now `unit` to match astropy style. A new `unit_str`
  attribute is a string derived dynamically from `unit`.

This renaming is also extended to the metrics.yaml schema.

Datum refactor
--------------

This includes a refactor of Datum with QuantityAttributeMixin.

This mixin provides common attributes for all classes that essentially
wrap a quantity.

Tests
-----

Tests are also improved significantly.

Documentation
-------------

Update documentation with new API design.