Add from_json classmethods to rebuild from JSON
This offers round-tripping objects -> json -> objects. Doing this can be
useful for unit tests of an analysis code against a set of known outputs
serialized to JSON using the validate_base api.

- Datum.from_json
- Specification.from_json
- Metric.from_json
- BlobBase.from_json
- MeasurementBase.from_json
- Job.from_json

Since BlobBase and MeasurementBase are abstract base classes, the
DeserializedBlob and DeserializedMeasurement classes are concrete
implementation for deserialization. They should work identically to the
original subclasses, except that they do no have additional methods and
non-serialized data. This approach lets us deserialize a JSON dataset
without needing the original Measurement and Blob classes to be
available.

To facilitate this, the serialization format for blob IDs in
measurement classes was changed to include the attribute names of the
blobs.