Fully support unitless int quantities
Turns out that when you make an integer into a unitless astropy quantity

q = 42 * u.Quantity('')

Then q.value will be a float. This isn't great. Upon reflection, it
seems that integers should always be unitless (because they're counts).
So there's no value in wrapping it in an integer. Thus, this commit
makes int a first-class type, alongside Quantity, str, bool and None.

This affects:

- Datum value
- Measurement parameters & extras (the measurement quantity itself should
  still be a Quantity or None).
- Specification parameters (the specification level should still be a
  quantity)