Use property and setter to wrap Measurement metric
The MeasurementBase abstract base class needs to ensure that measurement
classes have a metric instance. Previously this was enforced with an
abstract property. The downside was that the instance needed to exist
even before the measurement was instantiated, resulting in ugly
`metric = None` assignments in the measurement classes.

Using metric as a property that fronts self._metric gets around that
problem by throwing an exception at runtime if the metric was not
assigned.