Refactor Specification into ThresholdSpecification
This is a wide-ranging commit that effects the Specification data model
described in https://sqr-017.lsst.io

The main feature is that specification is an abstract base class now
that defines an interface for types of specifications. The one that
we've already been using is a threshold-like specifications, so that
functionality is refactored out into a ThresholdSpecification class.

Specifications have types (e.g., `threshold`), and also have a check()
method that tests a measurement against a specification.

When serializing, all specifications have 'name' and 'type' JSON fields.
Information particular to the type of specification is embedded in a
field named after the type itself. Specification subclasses provide this
value by implementing a _serialize_type method.

Other upgrades:

- __eq__ and __repr__ are implemented in ThresholdSpecification
- improved argument checking in ThresholdSpecification, raising
  TypeError otherwise.
- Use the Name class instead of a string for the specification's name.