Name raises TypeError from bad input during __init__
My reading is that ValueError is used for bad arguments to methods or
functions, but TypeError is used when a class cannot be constructed.
Thus I think TypeError is more Pythonic here.

In some cases Name.__init__ uses a static method for string parsing. In
this cases I catch their ValueError and re-raise a TypeError.