Add resolve_document method
- add SpecificationSet.resolve_document to resolve all inheritance in a
  YAML document.
- Add SpecificationResolutionError, which is our signal that a
  specification cannot be fully resolved.

resolve_document's job is to take a specification or specification
partial YAML *document* (that has been parsed into a dict already) and
resolves it's inheritance by merging information from base Specification
and SpecificationPartial objects that have already been parsed and
stored in the SpecificationSet instance.

This method can be used standalone to resolve an arbitrary new YAML
document, but it'll also be used by SpecificationSet itself to parse
and load specification definitions in YAML files.