Implement __ne__ methods
I've realized that the Python data model doesn't imply that __ne__ is
the logical negation of any implemented __eq__() method. Unit tests on
Python 2.7 ran into this issue (though testing under Python 3.5 was
unaffected, oddly).

This commit adds concrete implementation of __ne__ that are in fact the
logical negation of __eq__.

Affects classes:

- Name
- Measurement
- ThresholdSpecification
- Metric