Implement Job class for top-level data structure
Job contains measurements, metrics, and specifications, and allows the
full verification dataset to be serialized to JSON.

The Job class is responsible for serializing Blobs belonging to
measurements by collecting Blobs into a BlobSet and serializing that.

In the future, Job provenance will also be serialized through the Job
class.