Create Job.load_metrics_package classmethod
This is a convenient API for creating a Job object and pre-loading it
with metrics and specifications.

This API matches, and indeed composes, the load_metrics_package
classmethods of SpecificationSet and MetricSet.

I've also added a `subset` method load only metrics and specifications
for a single package. This might be useful in cases where a Job is being
created in a Task, and there's a desire to include the metric and
specification definitions, but no need to parse all of verify_metrics.