Add Metadata class to track job metadata
Metadata is a mapping type. It's attached to a Job's Job.meta attribute
and serializes to a 'meta' field in a job's JSON serialization.

Metadata is interesting because it can hold global meta in itself, but
it can also get measurement **annotations**. If a key is prefixed with a
metric name, then Metadata gets the metadata field from the
corresponding measurement's Measurement.notes attribute.

Internally this is implemented with collections.ChainMap, making
Metadata looks like a single `dict`, but really it is a head dict along
with the metadata dicts of each measurement.

This ChaimMap approach ensures that data is only stored in one place.
Annotations might be attached to a measurement before it's added to a
Job, so measurement annotations need to be stored in the measurement. At
the same time, all this metadata needs to be globally unified to support
specification metadata queries.

The downside is that simply key-value look-ups need to ensure that the
ChainMap is up to date, and that no new measurements have been added.

One way around this would be to add a ChainMap to MeasurementSet.
MeasurementSet knows when a new measurement is added, so it can be
proactive in maintaining its own ChainMap.