Add Measurement.notes attr and MeasurementNotes
Measurement.notes is a way of adding metadata directly to a measurement.
This metadata can be queries against by specifications (MetadataQuery)
to determine if a specification applies to the measurement.

Measurement.notes is a MeasurementNotes instance. MeasurementNotes
implements a mapping API, with the difference that it internally
namespaces all keys with the metric's name. This is done so that
measurement annotations can be added directly to a Job's metadata
mapping without conflicts.