Make a copy of spec when resolving it
If a specification has a base that can't be resolved, then an exception
is thrown in the resolve() method, indicating that it should be added to
a retry pile. But we're also modifying that specification in the
resolve_document() method. So when we actually do retry the resolution,
the specification may not be in a good state.

Doing a deep copy of the original specification in resolve_document()
solves all these things.