Add SpecificationSet.subset required_meta mode
The required_meta mode is the opposite of the regular metadata query.

Only specifications that have **all** the terms in ``required_metadata``
(and their term values match) are selected.  This is opposite to the
logic of the ``meta`` argument where a specification with an empty
metadata query is always selected, for example.

While the regular metadata query is correct for finding specifications
that apply to a job, the `required_meta` mode is useful for discovering
specifications.

At the level of Specification.query_metadata and MetadataQuery.__call__,
the `arg_driven` keyword argument activates this mode.

Added tests and clarified why each test case passes or fails to match.