More lenient ThresholdSpecification repr test
Doing the string match with a floating point Quantity is fraught --- and
created an issue for a numpy 1.14 upgrade. This substring matching makes
sure the that ThresholdSpecification.__repr__ is generally providing the
right info.

Note we can't test this repr by instantiating the object from the repr
string becaues the Astropy quantity __repr__ isn't instantiable.