Raise RuntimeError if replacing specs/partials
It seems that silently overwriting specifications or partials is leading
to hard-to-debug issues. This makes SpecificationSet raise a
RuntimeError exception if specifications or partials are being ingested
that replace previously-ingested specifications or partials of the same
name.

This also consolidates the act few setting the internals _specs and
_partials dicts to the __getitem__ method. Regular code shouldn't be
touching _specs or _partials.

Adjust tests to deal with new behavior