Update responses-based tests for new client.
The lsst.verify.squash client got updated in DM-13237 but we forgot to
check the tests with responses installed, which does finer-grained
testing of the HTTP client itself. This updates those tests to better
reflect the current client.

Perhaps we should add responses to the stack so that everyone can take
advantage of these tests?