Catch RuntimeErrors from getPackageDir
lsst.utils.getPackageDir normally emits a LookupError when it cannot
find a package, but in the case of a user passing a string
"verify_astrometry" in, it was emitting a RuntimeError instead. The
handling pattern is the same either way, so this addresses that rare
situation.

Not that the LookupError situation is already exercised in the unit
tests. I don't know if its worth try to force a RuntimeError situation
in the tests too for this.